+++ src/main/java/dto/RestaurantUpdateRequest.java
@@ -10,8 +10,8 @@ public class RestaurantUpdateRequest {
     private String address;
     private String phone;
     private String logoBase64;
-    private Double tax_fee;
-    private Double additional_fee;
+    private Integer tax_fee;
+    private Integer additional_fee;
 
     public RestaurantUpdateRequest() {
     }
